pb: skip unnecessary checks

+++ pb/utils.go
@@ -6,11 +6,11 @@ import (
 
 func AsTimestamp(t time.Time) *Timestamp {
 	return &Timestamp{
-		Seconds: int64(t.Unix()),
+		Seconds: t.Unix(),
 		Nanos:   int32(t.Nanosecond()),
 	}
 }
 
 func (t *Timestamp) AsTime() time.Time {
-	return time.Unix(int64(t.GetSeconds()), int64(t.GetNanos())).UTC()
+	return time.Unix(t.GetSeconds(), int64(t.GetNanos())).UTC()
 }
